fix(stargate): fail-closed bridge URL + mount the bridge into the Railway backend
The production Stargate page fell back to http://localhost:8100 when
NEXT_PUBLIC_STARGATE_BRIDGE_URL was unset, pointing novendor.ai visitors'
browsers at their own machines (ERR_CONNECTION_REFUSED, stuck "reconnecting").
Two-part fix (story #536):

Part A - frontend fail-closed: bridgeBaseUrl() returns the configured value,
the localhost fallback ONLY in next dev, else null. The hook creates no
EventSource when null and exposes `configured`; the console renders an explicit
"bridge not configured" diagnostic instead of silently failing.

Part B - deploy the bridge: the standalone bridge core moves into the backend
(app/services/stargate_bridge.py + app/routes/stargate_bridge.py) behind
STARGATE_BRIDGE_ENABLED (default off, capture mode in prod), so it ships with
Railway and the browser reaches it at a real HTTPS origin. signal_mapping and
the capture fixture move into backend/app (single source; the scripts dir keeps
re-export shims so producer/tests/demo_dryrun.ps1 are unchanged). Import purity
held (no app.config/app.observability) so `uvicorn bridge:app` still works on a
DB-less laptop. CORS via the existing ALLOWED_ORIGINS env (ops step). Endpoints
are unauthenticated GETs over synthetic fixture telemetry only (ADR 0003).

Evidence: backend 30 passed/1 skipped (incl. router-surface, 503-when-uninit,
default-off, wrapper smoke); frontend full suite 776 passed (new
stargateStream + StargateConsole tests); ruff/lint/build green; laptop wrapper
capture-mode health OK; fixture regenerates byte-identical at the new path;
flag-on SSE smoke against full app.main streamed 3 frames through the
logging+auth middleware stack.
