+++ docs/tips.md
@@ -3201,3 +3201,7 @@ nothing throws. Bonus diagnosis lesson re-confirmed: the same browser-agent revi
 WebGL error on a page with zero three.js imports (Factory ML) while simultaneously describing it
 rendering fully — agent-mode artifacts ride along with real findings; verify each against the code
 before fixing.
+
+### Public frontend deployments must never default to a localhost service URL outside development (2026-06-12)
+
+A `process.env.NEXT_PUBLIC_X || "http://localhost:8100"` fallback shipped to production: when the Vercel env var was unset, every novendor.ai visitor's browser tried to reach THEIR OWN machine (ERR_CONNECTION_REFUSED), and the page sat on "reconnecting". The fallback was right for `next dev` and wrong for a deployed page. Fix: resolve fail-closed — `return configured ?? (NODE_ENV === "development" ? localhost : null)` — and render an explicit "service URL not configured for this deployment" diagnostic when null instead of constructing the EventSource/fetch. Two load-bearing details: (1) read `NEXT_PUBLIC_*` and `NODE_ENV` by STATIC property access (`process.env.NEXT_PUBLIC_FOO`), never a dynamic `process.env[name]` — Next inlines the static form at build time and a dynamic lookup is `undefined` in the production bundle; (2) the test that locks it is the unconfigured render: jsdom has no EventSource, so asserting "zero EventSource constructions + diagnostic visible" both proves fail-closed AND that the page doesn't crash. The deeper fix was deploying the backing service (mount the standalone FastAPI bridge into the existing Railway backend behind a default-off flag, capture mode in prod) so the env var points at a real HTTPS origin — a public page needs a deployment contract, not a dev convenience. When a demo service is "standalone for the laptop", decide its production home before the page ships, or the localhost assumption leaks to real users.
